+++ Source/Core/Controls/AngleControl.cs
@@ -88,18 +88,18 @@ namespace CodeImp.DoomBuilder.Controls
 		// Redraw the control
 		private void AngleControl_Paint(object sender, PaintEventArgs e)
 		{
-			float rad = Angle2D.DegToRad(angle);
+			double rad = Angle2D.DegToRad(angle);
 			e.Graphics.PixelOffsetMode = PixelOffsetMode.HighQuality;
 			e.Graphics.InterpolationMode = InterpolationMode.High;
 			e.Graphics.SmoothingMode = SmoothingMode.HighQuality;
 			e.Graphics.Clear(this.BackColor);
 			Pen linepen = new Pen(SystemColors.ControlText, LINE_THICKNESS);
 			PointF start = new PointF(this.Size.Width * 0.5f, this.Size.Height * 0.5f);
-			float line_length = this.Size.Width * 0.26f;
+			double line_length = this.Size.Width * 0.26f;
 			if((rad >= 0) && (rad < 360))
 			{
-				PointF end = new PointF(start.X + (float)Math.Sin(rad + Angle2D.PIHALF) * line_length,
-										start.Y + (float)Math.Cos(rad + Angle2D.PIHALF) * line_length);
+				PointF end = new PointF((float)(start.X + Math.Sin(rad + Angle2D.PIHALF) * line_length),
+										(float)(start.Y + Math.Cos(rad + Angle2D.PIHALF) * line_length));
 				e.Graphics.DrawLine(linepen, start, end);
 			}
 			else
